Refactor regularizer class and test cases
