+++ tests/test_regularizers.py
@@ -2,165 +2,41 @@
 
 """Test that regularizers can be executed."""
 
-import logging
 import unittest
-from typing import Any, ClassVar, Dict, Optional, Type
+from typing import Dict
 
 import pytest
 import torch
 from torch.nn import functional
 
-from pykeen.datasets import Nations
-from pykeen.models import ConvKB, RESCAL, TransH
+from pykeen.models import ConvKB, TransH
 from pykeen.regularizers import (
     CombinedRegularizer, LpRegularizer, NoRegularizer, PowerSumRegularizer, Regularizer,
-    TransHRegularizer, _get_expected_norm,
+    TransHRegularizer,
 )
-from pykeen.triples import TriplesFactory
-from pykeen.typing import MappedTriples
-from pykeen.utils import resolve_device
+from pykeen.utils import get_expected_norm, resolve_device
+from tests import cases
 
 
-class _RegularizerTestCase:
-    """A test case for quickly defining common tests for regularizers."""
-
-    #: The batch size
-    batch_size: int
-    #: The triples factory
-    triples_factory: TriplesFactory
-    #: Class of regularizer to test
-    regularizer_cls: ClassVar[Type[Regularizer]]
-    #: The constructor parameters to pass to the regularizer
-    regularizer_kwargs: Optional[Dict[str, Any]] = None
-    #: The regularizer instance, initialized in setUp
-    regularizer: Regularizer
-    #: A positive batch
-    positive_batch: MappedTriples
-    #: The device
-    device: torch.device
-
-    def setUp(self) -> None:
-        """Set up the test case with a triples factory and model."""
-        self.generator = torch.random.manual_seed(seed=42)
-        self.batch_size = 16
-        self.triples_factory = Nations().training
-        self.device = resolve_device()
-        self.regularizer = self.regularizer_cls(
-            device=self.device,
-            **(self.regularizer_kwargs or {}),
-        )
-        self.positive_batch = self.triples_factory.mapped_triples[:self.batch_size, :].to(device=self.device)
-
-    def test_model(self) -> None:
-        """Test whether the regularizer can be passed to a model."""
-        # Use RESCAL as it regularizes multiple tensors of different shape.
-        model = RESCAL(
-            triples_factory=self.triples_factory,
-            regularizer=self.regularizer,
-        ).to(self.device)
-
-        # Check if regularizer is stored correctly.
-        self.assertEqual(model.regularizer, self.regularizer)
-
-        # Forward pass (should update regularizer)
-        model.score_hrt(hrt_batch=self.positive_batch)
-
-        # Call post_parameter_update (should reset regularizer)
-        model.post_parameter_update()
-
-        # Check if regularization term is reset
-        self.assertEqual(0., model.regularizer.term)
-
-    def test_reset(self) -> None:
-        """Test method `reset`."""
-        # Call method
-        self.regularizer.reset()
-
-        self.assertEqual(0., self.regularizer.regularization_term)
-
-    def test_update(self) -> None:
-        """Test method `update`."""
-        # Generate random tensors
-        a = torch.rand(self.batch_size, 10, device=self.device, generator=self.generator)
-        b = torch.rand(self.batch_size, 20, device=self.device, generator=self.generator)
-
-        # Call update
-        self.regularizer.update(a, b)
-
-        # check shape
-        self.assertEqual((1,), self.regularizer.term.shape)
-
-        # compute expected term
-        exp_penalties = torch.stack([self._expected_penalty(x) for x in (a, b)])
-        expected_term = torch.sum(exp_penalties).view(1) * self.regularizer.weight
-        assert expected_term.shape == (1,)
-
-        self.assertAlmostEqual(self.regularizer.term.item(), expected_term.item())
-
-    def test_forward(self) -> None:
-        """Test the regularizer's `forward` method."""
-        # Generate random tensor
-        x = torch.rand(self.batch_size, 10, generator=self.generator)
-
-        # calculate penalty
-        penalty = self.regularizer.forward(x=x)
-
-        # check shape
-        assert penalty.numel() == 1
-
-        # check value
-        expected_penalty = self._expected_penalty(x=x)
-        if expected_penalty is None:
-            logging.warning(f'{self.__class__.__name__} did not override `_expected_penalty`.')
-        else:
-            assert (expected_penalty == penalty).all()
-
-    def _expected_penalty(self, x: torch.FloatTensor) -> torch.FloatTensor:
-        """Compute expected penalty for given tensor."""
-        return None
-
-
-class NoRegularizerTest(_RegularizerTestCase, unittest.TestCase):
+class NoRegularizerTest(cases.RegularizerTestCase):
     """Test the empty regularizer."""
 
-    regularizer_cls = NoRegularizer
+    cls = NoRegularizer
 
     def _expected_penalty(self, x: torch.FloatTensor) -> torch.FloatTensor:  # noqa: D102
         return torch.zeros(1, device=x.device, dtype=x.dtype)
 
 
-class _LpRegularizerTest(_RegularizerTestCase):
-    """Common test for L_p regularizers."""
-
-    regularizer_cls = LpRegularizer
-
-    def _expected_penalty(self, x: torch.FloatTensor) -> torch.FloatTensor:  # noqa: D102
-        kwargs = self.regularizer_kwargs
-        if kwargs is None:
-            kwargs = {}
-        p = kwargs.get('p', self.regularizer.p)
-        value = x.norm(p=p, dim=-1).mean()
-        if kwargs.get('normalize', False):
-            dim = torch.as_tensor(x.shape[-1], dtype=torch.float, device=x.device)
-            if p == 2:
-                value = value / dim.sqrt()
-            elif p == 1:
-                value = value / dim
-            else:
-                raise NotImplementedError
-        return value
-
-
-class L1RegularizerTest(_LpRegularizerTest, unittest.TestCase):
+class L1RegularizerTest(cases.LpRegularizerTest):
     """Test an L_1 normed regularizer."""
 
-    regularizer_kwargs = {'p': 1}
+    kwargs = {'p': 1}
 
 
-class NormedL2RegularizerTest(_LpRegularizerTest, unittest.TestCase):
+class NormedL2RegularizerTest(cases.LpRegularizerTest):
     """Test an L_2 normed regularizer."""
 
-    regularizer_kwargs = {'normalize': True, 'p': 2}
+    kwargs = {'normalize': True, 'p': 2}
 
     @pytest.mark.slow
     def test_expected_norm(self):
@@ -168,7 +44,7 @@ class NormedL2RegularizerTest(_LpRegularizerTest, unittest.TestCase):
         n = 100
         for p in (1, 2, 3):
             for d in (2, 8, 64):
-                e_norm = _get_expected_norm(p=p, d=d)
+                e_norm = get_expected_norm(p=p, d=d)
                 norm = torch.randn(n, d).norm(p=p, dim=-1).numpy()
                 norm_mean = norm.mean()
                 norm_std = norm.std()
@@ -177,14 +53,14 @@ class NormedL2RegularizerTest(_LpRegularizerTest, unittest.TestCase):
 
         # test error is raised
         with pytest.raises(NotImplementedError):
-            _get_expected_norm(p=float('inf'), d=d)
+            get_expected_norm(p=float('inf'), d=d)
 
 
-class CombinedRegularizerTest(_RegularizerTestCase, unittest.TestCase):
+class CombinedRegularizerTest(cases.RegularizerTestCase):
     """Test the combined regularizer."""
 
-    regularizer_cls = CombinedRegularizer
-    regularizer_kwargs = {
+    cls = CombinedRegularizer
+    kwargs = {
         'regularizers': [
             LpRegularizer(weight=0.1, p=1, device=resolve_device()),
             LpRegularizer(weight=0.7, p=2, device=resolve_device()),
@@ -192,20 +68,20 @@ class CombinedRegularizerTest(_RegularizerTestCase, unittest.TestCase):
     }
 
     def _expected_penalty(self, x: torch.FloatTensor) -> torch.FloatTensor:  # noqa: D102
-        regularizers = self.regularizer_kwargs['regularizers']
+        regularizers = self.kwargs['regularizers']
         return sum(r.weight * r.forward(x) for r in regularizers) / sum(r.weight for r in regularizers)
 
 
-class PowerSumRegularizerTest(_RegularizerTestCase, unittest.TestCase):
+class PowerSumRegularizerTest(cases.RegularizerTestCase):
     """Test the power sum regularizer."""
 
-    regularizer_cls = PowerSumRegularizer
+    cls = PowerSumRegularizer
 
     def _expected_penalty(self, x: torch.FloatTensor) -> torch.FloatTensor:  # noqa: D102
-        kwargs = self.regularizer_kwargs
+        kwargs = self.kwargs
         if kwargs is None:
             kwargs = {}
-        p = kwargs.get('p', self.regularizer.p)
+        p = kwargs.get('p', self.instance.p)
         value = x.pow(p).sum(dim=-1).mean()
         if kwargs.get('normalize', False):
             value = value / x.shape[-1]
@@ -215,9 +91,11 @@ class PowerSumRegularizerTest(_RegularizerTestCase, unittest.TestCase):
 class TransHRegularizerTest(unittest.TestCase):
     """Test the TransH regularizer."""
 
+    cls = TransHRegularizer
     generator: torch.Generator
     device: torch.device
-    regularizer_kwargs: Dict
+
+    instance: Regularizer
     num_entities: int
     num_relations: int
     entities_weight: torch.Tensor
@@ -228,10 +106,10 @@ class TransHRegularizerTest(unittest.TestCase):
         """Set up the test case."""
         self.generator = torch.random.manual_seed(seed=42)
         self.device = resolve_device()
-        self.regularizer_kwargs = {'weight': .5, 'epsilon': 1e-5}
-        self.regularizer = TransHRegularizer(
+        self.kwargs = {'weight': .5, 'epsilon': 1e-5}
+        self.instance = self.cls(
             device=self.device,
-            **(self.regularizer_kwargs or {}),
+            **(self.kwargs or {}),
         )
         self.num_entities = 10
         self.num_relations = 5
@@ -243,7 +121,7 @@ class TransHRegularizerTest(unittest.TestCase):
         """Test update function of TransHRegularizer."""
         # Tests that exception will be thrown when more than or less than three tensors are passed
         with self.assertRaises(KeyError) as context:
-            self.regularizer.update(
+            self.instance.update(
                 self.entities_weight,
                 self.normal_vector_weight,
                 self.relations_weight,
@@ -251,22 +129,22 @@ class TransHRegularizerTest(unittest.TestCase):
             )
             self.assertTrue('Expects exactly three tensors' in context.exception)
 
-            self.regularizer.update(
+            self.instance.update(
                 self.entities_weight,
                 self.normal_vector_weight,
             )
             self.assertTrue('Expects exactly three tensors' in context.exception)
 
         # Test that regularization term is computed correctly
-        self.regularizer.update(self.entities_weight, self.normal_vector_weight, self.relations_weight)
+        self.instance.update(self.entities_weight, self.normal_vector_weight, self.relations_weight)
         expected_term = self._expected_penalty()
-        weight = self.regularizer_kwargs.get('weight')
-        self.assertAlmostEqual(self.regularizer.term.item(), weight * expected_term.item())
+        weight = self.kwargs.get('weight')
+        self.assertAlmostEqual(self.instance.term.item(), weight * expected_term.item())
 
     def _expected_penalty(self) -> torch.FloatTensor:  # noqa: D102
         # Entity soft constraint
         regularization_term = torch.sum(functional.relu(torch.norm(self.entities_weight, dim=-1) ** 2 - 1.0))
-        epsilon = self.regularizer_kwargs.get('epsilon')  #
+        epsilon = self.kwargs.get('epsilon')  #
 
         # Orthogonality soft constraint
         d_r_n = functional.normalize(self.relations_weight, dim=-1)
@@ -333,3 +211,11 @@ class TestOnlyUpdateOnce(unittest.TestCase):
         regularizer.reset()
         self.assertFalse(regularizer.updated)
         self.assertEqual(0.0, regularizer.regularization_term.item())
+
+
+class TestRegularizerTests(cases.TestsTestCase):
+    """Test all regularizers are tested."""
+
+    base_cls = Regularizer
+    base_test = cases.RegularizerTestCase
+    skip_cls = {TransHRegularizer}
